dev/alpha/requirement.txt
colorama==0.3.7
Django==1.10.4
djangorestframework==3.5.3
httpie==0.9.9
mysqlclient==1.3.10
Pygments==2.1.3
requests==2.12.4
psycopg2
pandas